+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## May 20th, 2024
+
+- Add allocation on login
+- Allocation connect
+- Allocation disconnect
+
 ## May 8th, 2024
 
 - Add Login command